[LibOS] Rework IPC transport layer
Old version was bugged, unstable, unnecessary complex and hard to
maintain.
After this PR other parts of code need not to worry about ipc
implementation details. As a bonus there is no 0.5 sleep on the program
exit.
IPC connections are now managed using two sets: incoming and outgoing.
The former is fully managed by a IPC worker thread, which handles new
connections and all incoming IPC messages. The latter is shared between
all threads, but encapsulated in the IPC module and the only thing you
need to send a IPC message is the destination process ID.

+++ LibOS/shim/include/shim_internal.h
@@ -302,7 +302,7 @@ static inline int clear_event(AEVENTTYPE* e) {
 #define REF_GET(ref)        __atomic_load_n(&(ref).counter, __ATOMIC_SEQ_CST)
 #define REF_SET(ref, count) __atomic_store_n(&(ref).counter, count, __ATOMIC_SEQ_CST);
 
-static inline int __ref_inc(REFTYPE* ref) {
+static inline int64_t __ref_inc(REFTYPE* ref) {
     int64_t _c;
     do {
         _c = __atomic_load_n(&ref->counter, __ATOMIC_SEQ_CST);
@@ -314,7 +314,7 @@ static inline int __ref_inc(REFTYPE* ref) {
 
 #define REF_INC(ref) __ref_inc(&(ref))
 
-static inline int __ref_dec(REFTYPE* ref) {
+static inline int64_t __ref_dec(REFTYPE* ref) {
     int64_t _c;
     do {
         _c = __atomic_load_n(&ref->counter, __ATOMIC_SEQ_CST);
